Remove common lucene dependency, refactor SignatureBuilder out of common
