+++ kifi-backend/modules/common/app/com/keepit/rover/document/utils/Signature.scala
@@ -1,11 +1,7 @@
 package com.keepit.rover.document.utils
 
-import java.io.StringReader
 import java.util.Arrays
 import javax.xml.bind.DatatypeConverter._
-
-import org.apache.lucene.analysis.standard.StandardTokenizer
-import org.apache.lucene.analysis.tokenattributes.CharTermAttribute
 
 import scala.math._
 
@@ -111,19 +107,3 @@ object Signature {
   }
 }
 
-class SignatureBuilder(windowSize: Int = 20) extends Signature.Builder(windowSize) {
-
-  override protected def tokenize(text: String)(addTerm: (Array[Char], Int) => Unit): Unit = {
-    val ts = new StandardTokenizer(new StringReader(text))
-    val termAttr = ts.getAttribute(classOf[CharTermAttribute])
-
-    try {
-      ts.reset()
-      while (ts.incrementToken()) addTerm(termAttr.buffer(), termAttr.length())
-      ts.end()
-    } finally {
-      ts.close()
-    }
-  }
-
-}
