+++ kifi-backend/project/Global.scala
@@ -82,8 +82,7 @@ object Global {
     "com.kifi" %% "json-annotation" % "0.2",
     "com.mchange" % "c3p0" % "0.9.5-pre8", // todo(andrew): remove from common when C3P0 plugin is in sqldb
     "org.im4java" % "im4java" % "1.4.0", //todo(eishay): means that all services get that. not sure that's this is best
-    "org.apache.tika" % "tika-parsers" % "1.5", // todo(Léo): remove when Scraper is gone and Tika is in Rover,
-    "org.apache.lucene" % "lucene-analyzers-common" % "4.10.2" // todo(Léo): remove when Scraper is gone and Signature is in Rover,
+    "org.apache.tika" % "tika-parsers" % "1.5" // todo(Léo): remove when Scraper is gone and Tika is in Rover,
   ) map (_.excludeAll(
     ExclusionRule(organization = "javax.jms"),
     ExclusionRule(organization = "com.sun.jdmk"),
